refactor(auth-service): rename structure to improve clarity and maintainability
