+++ src/app/core/services/auth/authService/auth-role.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { Injectable } from '@angular/core';
-import { AuthTokenService } from './auth-token.service';
+import { AuthTokenService } from './auth-token';
 
 @Injectable({
   providedIn: 'root',
